Torch animation which controls the collider activation was added to the right hand of the player prefab. When the collider is active, it triggers the damage when another non-trigger collider is detected. The ennemy MUST have the tag "Ennemy" and the script GenericHealth
