+++ Snowy-Path/Assets/Animations/Torch.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 769c6bb4dd9952d4bb5afb3c1bc552d6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
