+++ Snowy-Path/Assets/Animations/Torch/Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb1fe6bb9757c7247b1a5518fa756d2d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
