+++ Snowy-Path/Assets/Animations/Torch/Sweep.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e08e5586e7da39640be3f3cb644e4236
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
