+++ Snowy-Path/Assets/Animations/Torch/Torch.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff1b9c507480c004e9e3640806502baf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
