+++ Snowy-Path/Assets/Inputs/PlayerActionInputs.inputactions
@@ -184,8 +184,8 @@
                 },
                 {
                     "name": "",
-                    "id": "6226a900-cc75-4afc-95dc-5cfc8c80ccfe",
-                    "path": "<Mouse>/rightButton",
+                    "id": "b6ff6111-0112-41cd-8a5a-bd48c1d6a43e",
+                    "path": "<Mouse>/leftButton",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
@@ -217,8 +217,8 @@
                 },
                 {
                     "name": "",
-                    "id": "312961ab-4c32-4877-8e57-54b1ccbda9c9",
-                    "path": "<Gamepad>/rightTrigger",
+                    "id": "6226a900-cc75-4afc-95dc-5cfc8c80ccfe",
+                    "path": "<Mouse>/rightButton",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
@@ -228,8 +228,8 @@
                 },
                 {
                     "name": "",
-                    "id": "b6ff6111-0112-41cd-8a5a-bd48c1d6a43e",
-                    "path": "<Mouse>/leftButton",
+                    "id": "312961ab-4c32-4877-8e57-54b1ccbda9c9",
+                    "path": "<Gamepad>/rightTrigger",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
