+++ Snowy-Path/Assets/Scripts/Tool/Torch.cs.meta
