+++ Snowy-Path/ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - NorthPole
   - HeatSource
   - WeatherZone
+  - Ennemy
   layers:
   - Default
   - TransparentFX
